re-committing all
