+++ lib/views/enrollmentPage/courses_list_view.dart
@@ -16,6 +16,9 @@ class CoursesListView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    double deviceWidth = MediaQuery.of(context).size.width;
+    double deviceHeight = MediaQuery.of(context).size.height;
+
     return Column(
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
@@ -24,30 +27,46 @@ class CoursesListView extends StatelessWidget {
           "COURSE OFFERINGS",
           style: TextStyle(fontSize: 18, fontWeight: FontWeight.w600),
         ),
-        const SizedBox(height: 10),
+        const SizedBox(height: 20),
         Expanded(
           child: ListView.builder(
             itemCount: courses.length,
             itemBuilder: (context, index) {
               final course = courses.elementAt(index);
-              return ListTile(
-                title: Text(
-                  course.courseName,
-                  maxLines: 1,
-                  softWrap: true,
-                  overflow: TextOverflow.ellipsis,
-                ),
-                trailing: IconButton(
-                  onPressed: () async {
-                    onEnrollCourse(course);
-                  },
-                  icon: const Icon(Icons.add_card_rounded),
-                ),
+              return DataTable(
+                dataRowHeight: deviceHeight * 0.1,
+                headingRowHeight: deviceHeight * 0,
+                columnSpacing: deviceWidth * 0.10,
+                columns: const [
+                  DataColumn(label: Text('Course Code', maxLines: 1,
+                    softWrap: true,
+                    overflow: TextOverflow.ellipsis,
+                  )),
+                  DataColumn(label: Text('Schedule', maxLines: 1,
+                    softWrap: true,
+                    overflow: TextOverflow.ellipsis,)),
+                  DataColumn(label: Text('+')),
+                ],
+                rows: [
+                  DataRow(cells: [
+                    DataCell(Text(course.courseCode)),
+                    DataCell(Center(child: Text(course.schedule.toString()))),
+                    DataCell(
+                      IconButton(
+                        onPressed: () async {
+                          onEnrollCourse(course);
+                        },
+                        icon: const Icon(Icons.add_card_rounded),
+                      ),
+                    ),
+                  ]),
+                ],
               );
             },
           ),
         ),
       ],
     );
+
   }
 }
